refactor: split net.js
