+++ server/src/objects/gas.ts
@@ -3,15 +3,6 @@ import { GameConfig } from "../../../shared/gameConfig";
 import { math } from "../../../shared/utils/math";
 import { type Game } from "../game";
 
-export interface GasData {
-    mode: number
-    duration: number
-    posOld: Vec2
-    posNew: Vec2
-    radOld: number
-    radNew: number
-}
-
 const GasMode = GameConfig.GasMode;
 export class Gas {
     stage = 0;
