+++ shared/msgs/emoteMsg.ts
@@ -1,0 +1,23 @@
+import { v2 } from "../utils/v2";
+import { AbstractMsg, BitStream } from "../net";
+
+
+export class EmoteMsg extends AbstractMsg {
+    pos = v2.create(0, 0);
+    type = "";
+    isPing = false;
+
+    override serialize(s: BitStream) {
+        s.writeVec(this.pos, 0, 0, 1024, 1024, 16);
+        s.writeGameType(this.type);
+        s.writeBoolean(this.isPing);
+        s.writeBits(0, 5);
+    }
+
+    override deserialize(s: BitStream) {
+        this.pos = s.readVec(0, 0, 1024, 1024, 16);
+        this.type = s.readGameType();
+        this.isPing = s.readBoolean();
+        s.readBits(5);
+    }
+}
