--- shared/netTypings.d.ts
@@ -1,5 +1,0 @@
-import { type BitStream } from "./net";
-
-export interface Msg {
-    serialize: (s: BitStream) => void
-}
